% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(n, a, b)
}
\arguments{
\item{n}{parameter in Bernoulli distribution}

\item{a}{parameter in Beta distribution}

\item{b}{parameter in Beta distribution}
}
\value{
a two-dimensional chain of size 10000
}
\description{
A Gibbs sampler to generate a chain with target joint density
}
\examples{
\dontrun{
sampleC = gibbsC(50, 30, 40)
plot(sampleC, cex = 0.5)
}
}
