% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JackAfterBoot.R
\name{JackAfterBoot}
\alias{JackAfterBoot}
\title{Jackknife-after-Bootstrap}
\usage{
JackAfterBoot(data, func = NULL, B)
}
\arguments{
\item{data}{the data as a vector}

\item{func}{the function which applied to data}

\item{B}{the number of replicates in Bootstrap}
}
\value{
the bias and standard deviation estimate for both Bootstrap and Jackknife-after-Bootstrap methods
}
\description{
Calculate the bias and standard deviation estimate for each leave-one-out Bootstrap sample.
}
\examples{
\dontrun{
data("cars")
data <- cars$speed
JackAfterBoot(data, mean, 1000)
JackAfterBoot(data, median, 1000)
}
}
