\name{StatComp21062-package}
\alias{StatComp21062-package}
\alias{StatComp21062}
\docType{package}
\title{
  Final work for the 'Statistical Computing' course
}
\description{
  Several R and Rcpp functions and all the homeworks are included in this package.
}
\author{
Yimeng Sun
Maintainer: Yimeng Sun <sunyimeng21@mail.ustc.edu.cn>
}
\keyword{ package }
